+++ src/beam/lib/SoundSourceLocalizer.h
@@ -15,19 +15,19 @@ class SoundSourceLocalizer {
 public:
 	SoundSourceLocalizer();
 	~SoundSourceLocalizer();
-	void init(float sample_rate, int frame_size);
-	void process(const std::vector<std::vector<std::complex<float> > >& input_channels, float* p_angle, float* p_weight);
+	void init(double sample_rate, int frame_size);
+	void process(const std::vector<std::vector<std::complex<double> > >& input_channels, double* p_angle, double* p_weight);
 
 private:
 	// sythetic data.
-	std::unique_ptr<float[]> m_delta[MAX_MICROPHONES - 1][NUM_ANGLES];
-	float m_angle[NUM_ANGLES];
+	std::unique_ptr<double[]> m_delta[MAX_MICROPHONES - 1][NUM_ANGLES];
+	double m_angle[NUM_ANGLES];
 	int m_start_bin;
 	int m_end_bin;
 	int m_meas_bins;
 	// common.
 	MicrophoneArray m_mic_array;
-	float m_sample_rate;
+	double m_sample_rate;
 	int m_frame_size;
 };
